Fix building on ppc64
The type definitions on ppc64 has two forms, the older one that uses long
for 64-bit types and the newer one that uses long long for 64-bit types.

The long long version is what other architectures use and what libbpfgo
expects in e.g. bpfMapCreateOptsToC.

All that's required to enable the latter ABI is to define
__SANE_USERSPACE_TYPES__ when including the type headers.